+++ src/screens/diary/DiaryEditorScreen.tsx
@@ -14,102 +14,17 @@ import {DiaryDatePicker} from '@src/components/diary/editor/date/DiaryDatePicker
 import {DiaryLocationPicker} from '@src/components/diary/editor/location/DiaryLocationPicker';
 import {DiaryFeelingPicker} from '@src/components/diary/editor/feeling/DiaryFeelingPicker';
 import {BottomSheetModalProvider} from '@gorhom/bottom-sheet';
-import {
-  PlaceholderBridge,
-  RichText,
-  TenTapStartKit,
-  Toolbar,
-  useEditorBridge,
-  DEFAULT_TOOLBAR_ITEMS,
-  useEditorContent,
-} from '@10play/tentap-editor';
-import {launchImageLibrary} from 'react-native-image-picker';
-import {Controller, FormProvider, useForm} from 'react-hook-form';
-import {DiarySchema} from '@src/libs/zod/DiarySchema';
-import {zodResolver} from '@hookform/resolvers/zod';
-import {Diary} from '@src/types/diary';
+import {RichText, Toolbar, DEFAULT_TOOLBAR_ITEMS} from '@10play/tentap-editor';
+import {Controller, FormProvider} from 'react-hook-form';
 import {useNavigation} from '@react-navigation/native';
 import {NavigationProps} from '@src/types/navigation';
 import {DiaryRegisterButton} from '@src/components/diary/editor/DiaryRegisterButton';
-import {BACKEND_URL} from '@env';
-import EncryptedStorage from 'react-native-encrypted-storage';
-import {useMutation} from '@tanstack/react-query';
 import {COLOR} from '@src/constants/color';
+import {useDiaryEditor} from '@src/hooks/diary/useDiaryEditor';
 
 export const DiaryEditorScreen = () => {
+  const {methods, editor, imageMutation, handleImageUpload} = useDiaryEditor();
   const navigation = useNavigation<NavigationProps>();
-  const editor = useEditorBridge({
-    avoidIosKeyboard: true,
-    initialContent: '',
-    bridgeExtensions: [
-      ...TenTapStartKit,
-      PlaceholderBridge.configureExtension({
-        placeholder:
-          '여행은 어땠나요? 자유롭게 기록하고 싶은 것들을 작성해보세요.',
-      }),
-    ],
-  });
-  const content = useEditorContent(editor, {type: 'html'});
-  const methods = useForm<Diary>({
-    resolver: zodResolver(DiarySchema),
-    defaultValues: {
-      title: '',
-      startDate: null,
-      endDate: null,
-      location: null,
-      feeling: null,
-      content: content,
-      image: null,
-    },
-    mode: 'onChange',
-  });
-
-  const imageMutation = useMutation({
-    mutationFn: async (image: {
-      fileName: string;
-      type: string;
-      uri: string;
-    }) => {
-      const accessToken = await EncryptedStorage.getItem('access_token');
-      const formData = new FormData();
-      formData.append('image', {
-        name: image.fileName,
-        type: image.type,
-        uri: image.uri,
-      });
-      formData.append('type', 'DIARY');
-
-      const response = await fetch(`${BACKEND_URL}/api/image`, {
-        method: 'POST',
-        headers: {
-          Cookie: `access_token=${accessToken}`,
-        },
-        body: formData,
-      });
-
-      if (!response.ok) {
-        throw new Error('Failed to upload an image.');
-      }
-
-      return await response.json();
-    },
-    onSuccess: (data: {fileUrl: string}) => {
-      methods.setValue('image', data.fileUrl);
-    },
-    throwOnError: true,
-  });
-
-  const handleImageUpload = () => {
-    launchImageLibrary({mediaType: 'photo'}, response => {
-      if (response.assets) {
-        imageMutation.mutate({
-          fileName: response.assets[0].fileName!,
-          type: response.assets[0].type!,
-          uri: response.assets[0].uri!,
-        });
-      }
-    });
-  };
 
   useEffect(() => {
     navigation.setOptions({
